pyproject.toml
[tool.poetry]
name = "ted-z"
version = "2.2.1"
description = "Just another Discord bot written in Python! (this is the last revision i swear)"
authors = ["Alex The JPEG <alexthejpeg@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
hikari-lightbulb = "^2.3.1"
hikari-miru = "^2.0.1"
requests = "^2.28.1"
libretranslatepy = "^2.1.3"
pillow = "^9.4.0"
beautifulsoup4 = "^4.11.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
ruff = "^0.0.236"
isort = "^5.11.4"

[tool.black]
line-length = 99
preview = true

[tool.ruff]
select = ["E", "F", "C", "I", "ANN", "B", "C4", "SIM", "PTH"]
ignore = ["ANN002", "ANN003", "ANN101"]
line-length = 99
target-version = "py311"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
